Define sparse tensors with Sparse*Tensor naming for better errors.
Previously, when you got a type error involving a tensor,
you got the following message:

    AttributeError: type object 'FloatTensor' has no attribute 'foo'

This error would be the same whether or not you had a sparse or
regular tensor.

With this commit, by defining the sparse tensor classes with sparse
in their name, and then creating aliases, means that now we will
see 'SparseFloatTensor' if there is an error.

Signed-off-by: Edward Z. Yang <ezyang@fb.com>