Hilarious new codegen on existing IR.
Ditch PExpr, instead use Expr. The hilarity lies in the reuse of
the same IR for both scalar and vector computation.  It is not
too surprising but there is now partiality: the interpreter
doesn't know how to run scalar ops (in principle it could be taught),
while the CUDA printer doesn't know how to handle vector ops (which
is what you'd expect.)

Signed-off-by: Edward Z. Yang <ezyang@fb.com>